// frontend/src/api/models/IcdCodeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IcdCodeResponse
 */
export interface IcdCodeResponse {
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    code?: string;
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    chapter?: string;
    /**
     * 
     * @type {string}
     * @memberof IcdCodeResponse
     */
    icdVersion?: string;
}

/**
 * Check if a given object implements the IcdCodeResponse interface.
 */
export function instanceOfIcdCodeResponse(value: object): value is IcdCodeResponse {
    return true;
}

export function IcdCodeResponseFromJSON(json: any): IcdCodeResponse {
    return IcdCodeResponseFromJSONTyped(json, false);
}

export function IcdCodeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): IcdCodeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'code': json['code'] == null ? undefined : json['code'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'chapter': json['chapter'] == null ? undefined : json['chapter'],
        'icdVersion': json['icdVersion'] == null ? undefined : json['icdVersion'],
    };
}

export function IcdCodeResponseToJSON(json: any): IcdCodeResponse {
    return IcdCodeResponseToJSONTyped(json, false);
}

export function IcdCodeResponseToJSONTyped(value?: IcdCodeResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'code': value['code'],
        'name': value['name'],
        'description': value['description'],
        'chapter': value['chapter'],
        'icdVersion': value['icdVersion'],
    };
}